fix plate update

+++ BackEnd/src/main/java/com/proyectointegradorequipo3/proyectointegradorEquipo3/services/impl/PlateServiceImpl.java
@@ -85,15 +85,25 @@ public class PlateServiceImpl implements IPlateService {
         try {
             PlateDto plateDto = searchPlateById(id);
 
-            plateDto.setName(request.getName());
-            plateDto.setType(request.getType());
-            plateDto.setDescription(request.getDescription());
+            if (request.getName() != null) {
+                plateDto.setName(request.getName());
+            }
 
-            s3Service.deleteObject(plateDto.getImage());
+            if (request.getType() != null) {
+                plateDto.setType(request.getType());
+            }
+
+            if (request.getDescription() != null) {
+                plateDto.setDescription(request.getDescription());
+            }
 
             MultipartFile newImage = request.getImage();
-            String newImageUrl = s3Service.putObject(newImage);
-            plateDto.setImage(newImageUrl);
+            if (newImage != null && !newImage.isEmpty()) {
+                s3Service.deleteObject(plateDto.getImage());
+
+                String newImageUrl = s3Service.putObject(newImage);
+                plateDto.setImage(newImageUrl);
+            }
 
             Plate plate = mapper.map(plateDto, Plate.class);
             plate.setId(id);
@@ -102,8 +112,6 @@ public class PlateServiceImpl implements IPlateService {
             throw new Exception(e.getMessage());
         }
     }
-
-
 
     //===================Delete===================//
